# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GameInCome",
        "description": "Description of API accessor methods for GameInCome",
        "contact": {
            "name": "Support",
            "email": "karlossan@me.com"
        },
        "license": {
            "name": "License 1.0 (Demo)",
            "url": "L5_SWAGGER_CONST_HOST"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://game-back.test/api/v1",
            "description": "Production API Server"
        },
        {
            "url": "https://game-back.karlossan.com/api/v1",
            "description": "Stage API Server"
        }
    ],
    "paths": {
...
        "/personal-area/purchases/{itemType}": {
            "get": {
                "tags": [
                    "Personal Area: Purchases"
                ],
                "summary": "get purchases",
                "operationId": "purchases",
                "parameters": [
                    {
                        "name": "itemType",
                        "in": "path",
                        "description": "games / software / swiches / cases",
                        "required": true,
                        "example": "games"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Недостаточно прав для выполнения запроса"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/personal-area/purchases/{itemType}/{keyId}": {
            "get": {
                "tags": [
                    "Personal Area: Purchases"
                ],
                "summary": "get purchase",
                "operationId": "purchase",
                "parameters": [
                    {
                        "name": "itemType",
                        "in": "path",
                        "description": "games / software / swiches / cases",
                        "required": true,
                        "example": "games"
                    },
                    {
                        "name": "keyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KeyPurchaseResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Недостаточно прав для выполнения запроса"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/personal-area/sales/{itemType}": {
            "get": {
                "tags": [
                    "Personal Area: Sales"
                ],
                "summary": "get items",
                "operationId": "sales",
                "parameters": [
                    {
                        "name": "itemType",
                        "in": "path",
                        "description": "games / software / swiches / cases",
                        "required": true,
                        "example": "games"
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "bought / archived",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SaleResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Недостаточно прав для выполнения запроса"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Personal Area: Sales"
                ],
                "summary": "TODO create items",
                "operationId": "sales_create",
                "parameters": [
                    {
                        "name": "itemType",
                        "in": "path",
                        "description": "games / software / swiches / cases",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Error: Bad Request"
                    },
                    "401": {
                        "description": "Error: Unauthenticated"
                    },
                    "403": {
                        "description": "Error: Forbidden / Недостаточно прав для выполнения запроса"
                    },
                    "404": {
                        "description": "Error: Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/personal-area/sales/{itemType}/{itemId}": {
            "put": {
                "tags": [
                    "Personal Area: Sales"
                ],
                "summary": "update items",
                "operationId": "sales_update",
                "parameters": [
                    {
                        "name": "itemType",
                        "in": "path",
                        "description": "games / software / swiches / cases",
                        "required": true
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SalesUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Personal Area: Sales"
                ],
                "summary": "delete items",
                "operationId": "sales_delete",
                "parameters": [
                    {
                        "name": "itemType",
                        "in": "path",
                        "description": "games / software / swiches / cases",
                        "required": true
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SalesUpdateRequest": {
                "properties": {
                    "archive": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "KeyPurchaseResource": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "itemName": {
                                    "type": "string"
                                },
                                "itemBackground": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "number"
                                },
                                "keyId": {
                                    "type": "string"
                                },
                                "key": {
                                    "type": "string"
                                },
                                "sellerId": {
                                    "type": "string"
                                },
                                "sellerName": {
                                    "type": "string"
                                },
                                "sellerSurname": {
                                    "type": "string"
                                },
                                "sellerNickname": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "type": "integer",
                        "example": "200"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "KeyPurchaseResource"
                }
            },
            "PurchaseResource": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "purchaseId": {
                                    "type": "string"
                                },
                                "keyId": {
                                    "type": "string"
                                },
                                "key": {
                                    "type": "string"
                                },
                                "itemName": {
                                    "type": "string"
                                },
                                "itemBackground": {
                                    "type": "string"
                                },
                                "itemPlatform": {
                                    "type": "string"
                                },
                                "createdAt": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "type": "integer",
                        "example": "200"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PurchaseResource"
                }
            },
            "SaleResource": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "itemId": {
                                    "type": "string"
                                },
                                "itemName": {
                                    "type": "string"
                                },
                                "itemService": {
                                    "type": "string"
                                },
                                "platformId": {
                                    "type": "string"
                                },
                                "itemHeaderImage": {
                                    "type": "string"
                                },
                                "oldPrice": {
                                    "type": "number"
                                },
                                "newPrice": {
                                    "type": "number"
                                },
                                "sale": {
                                    "type": "number"
                                },
                                "keyId": {
                                    "type": "string"
                                },
                                "itemCreatedAt": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "status": {
                        "type": "integer",
                        "example": "200"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LoginResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer 35|a7aabf6c23d1...)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}